+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2019-02-10
+
+### Fixed
+
+- Bug that crashed delator if the user had not created a local database
+- Bug that crashed delator when A record lookups were requested on locally stored subdomains
+- Minor documentation fixes and updates
+- Miscellaneous fixes and improvements
+
 ## [1.2.0] - 2019-01-06
 
 ### Added